Cache clearing removed

+++ did/experiments/augmentation/params_search.py
@@ -1,3 +1,4 @@
+import os
 from time import time
 import itertools
 import numpy as np
@@ -17,8 +18,9 @@ if __name__ == "__main__":
         'way': [5],
         'train_shot': [1, 3, 5, 10],
         'test_shot': [1],
-        'x_dim': [28, 40, 50, 60, 80, 105],
-        'z_dim': [100, 200, 300, 500, 600, 784, 1000],
+        # ATTENTION: Due to the cached nature of dataloader this parameter should be set in signle value per run
+        'x_dim': [28],
+        'z_dim': [50, 100, 200, 300, 500, 600, 784, 1000],
         'optimizer': ['adam'],
         'lr': [0.01, 0.001, 0.0005],
         'initialization': ['orthogonal', 'xavier_normal', 'xavier_uniform'],
@@ -36,10 +38,11 @@ if __name__ == "__main__":
     keys, values = zip(*configs.items())
     param_grid = [dict(zip(keys, v)) for v in itertools.product(*values)]
 
-    # Create empty resulting file
-    res_path = "did/experiments/augmentation/results.csv"
-    df = pd.DataFrame(columns=configs.keys())
-    df.to_csv(res_path, index=False)
+    # Create resulting file if necessary
+    res_path = "did/experiments/base/results.csv"
+    if not os.path.exists(res_path):
+        df = pd.DataFrame(columns=configs.keys())
+        df.to_csv(res_path, index=False)
 
     conf_durations = []
     for i, param in enumerate(param_grid):
